fix code review

+++ features/merchant/gold_merchant_common/src/main/java/com/tokopedia/gm/common/domain/interactor/GetShopInfoPeriodUseCase.kt
@@ -54,7 +54,7 @@ class GetShopInfoPeriodUseCase @Inject constructor(
 
         @JvmStatic
         fun createParams(shopID: Long): RequestParams = RequestParams.create().apply {
-            putLong(SHOP_ID, shopID.toLong())
+            putLong(SHOP_ID, shopID)
         }
     }
